wip, starting findBy
