+++ packages/test-harness/src/findByLgId/index.ts
@@ -1,0 +1,14 @@
+import { makeSingleQuery } from '../utils';
+
+/**
+ * TODO: Add description here
+ * @param testId
+ * @returns
+ */
+export const findByLgId = (testId: string) => {
+  const findBy = makeSingleQuery();
+
+  return waitFor(() => {
+    return findBy(testId);
+  });
+};
