+++ packages/test-harness/src/getByLgId/index.ts
@@ -1,4 +1,4 @@
-import { getDocument, getError } from '../utils';
+import { makeSingleQuery } from '../utils';
 
 /**
  * TODO: Add description here
@@ -6,19 +6,6 @@ import { getDocument, getError } from '../utils';
  * @returns
  */
 export const getByLgId = (testId: string) => {
-  const container = getDocument();
-  const dataString = `[data-lgid="${testId}"]`;
-  const element = container.querySelectorAll(dataString);
-
-  // TODO: console the container
-  if (!element.length) {
-    throw getError(`Unable to find an element by ${dataString}`);
-  }
-
-  // TODO: console the elements
-  if (element.length > 1) {
-    throw getError(`Found multiple elements by ${dataString}`);
-  }
-
-  return element[0];
+  const getBy = makeSingleQuery();
+  return getBy(testId);
 };
