+++ packages/test-harness/src/utils/index.ts
@@ -1,2 +1,3 @@
 export { getDocument } from './getDocument';
 export { getError } from './getError';
+export { makeSingleQuery } from './makeSingleQuery';
