+++ packages/test-harness/src/utils/waitFor/index.tsx
@@ -1,0 +1,3 @@
+// https://github.com/testing-library/dom-testing-library/blob/main/src/wait-for.js#L18
+
+const waitFor = () => {};
